Use BigDecimal instead of Double inside FxResponse

+++ src/main/java/com/galvanize/prodman/model/FxResponse.java
@@ -3,13 +3,14 @@ package com.galvanize.prodman.model;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Getter
 @Setter
 public class FxResponse {
 
-    private Map<String, Double> quotes;
+    private Map<String, BigDecimal> quotes;
 
     @Override
     public String toString() {
